Fixed lighting bug.
The shader program did not bind to the "Lights" uniform block but rather
tried to bind to "Lights[0]", "Lights[1]", and so on (as was needed in an
earlier implementation). This somehow still worked on Intel integrated
graphics card, but NVIDIA's card (correctly) gave some random lighting results.

Fixed: The mouse movement is now not relative to the delta time, and thus
gives more appropriate and consistent camera rotating speed.