Some minor adjustment to the shader.
The light attenuation constant is now less (now 0.5, earlier 1.0), which
allows brighter light closer to the light source.

Possible fix to a light problem: lights were not ordered in any way,
meaning some light that would have more influence to an object might not
be in the final MAX_LIGHTS lights. Now directional lights are inserted to
the front of the lights. This is not perfect, and will still fail in
situations where point lights should be ordered.