Improvements on pbs shader.
Added possibility to override materials by the material name exported to
the scene file. This feature is not used currently as it would override
materials for all objects as there is no way currently to select if the
user wants to override the material or not (in blender).

Some fixes in the GameObject, components were not deleted in the
destructor.