Improved image based lighting (environment mapping).
Material precalculates a look-up-table (texture) for environment mapping.
The LUT contains information of the specular term (scale, bias) based on
roughness (y coord.) and NoV (x coord.). This is used in the shader to
simplify the specular ibl calculations.

The LUT data is saved to "assets/env_brdf.data" for reuse. A debug preview
of the data is saved as "assets/env_brdf.tga".

Seamless cube maps are enabled allowing for better environment mapping.

Added 16 bit (normalized unsigned short) texure formats to Texture.