Added a puddle (a plane over a pit) to the scene.
water.frag: a water shader. Uses only env map for reflections. Normals are
calculated to be small waves.

pbs_ground.frag: added an effect to fake absorption by the water in the
puddle. This means all pits are "puddles". Improvement: the ground shader
now takes two samples from the albedo and normal maps and lerps them to
get rid of most of the repetition.

Fixed: uniform buffers now can correctly handle cases where the uniform
block has e.g. vec3 and float right next to each other. These need to be
added to the uniforms with no padding as the shader considers them to take
the space of a single vec4.